refactor management client skeleton
