+++ pkg/rabbitmq/management/types/main.go
@@ -1,0 +1,6 @@
+package types
+
+type ErrorResponse struct {
+	Error  string `json:"error"`
+	Reason string `json:"reason"`
+}
